ss-10: * refactor code to project structure
