+++ frontend/src/bundles/common/components/charts/line-chart/constants/constants.ts
