+++ frontend/src/bundles/common/components/charts/line-chart/line.type.ts
@@ -1,6 +1,3 @@
 type DataObject = { date: string | number; value: number };
-type ChartProperties = {
-    dataArr: DataObject[];
-};
 
-export { type ChartProperties, type DataObject };
+export { type DataObject };
