+++ frontend/src/bundles/common/components/charts/line-chart/styles.module.scss
@@ -1,4 +1,4 @@
 .chart {
-    height: 400px;
     width: 100%;
+    height: 400px;
 }
